[plugin.video.arteplussept] 1.5.1
