+++ plugin.video.arteplussept/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.arteplussept" name="Arte +7" version="1.4.5" provider-name="bmf, thomas-ernest">
+<addon id="plugin.video.arteplussept" name="Arte +7" version="1.5.1" provider-name="bmf, thomas-ernest">
     <!-- https://kodi.wiki/view/Addon.xml -->
     <requires>
         <import addon="xbmc.python" version="3.0.0"/>
@@ -55,18 +55,9 @@ https://github.com/thomas-ernest/plugin.video.arteplussept
         <license>MIT</license>
         <website>https://www.arte.tv/fr/</website>
         <source>https://github.com/thomas-ernest/plugin.video.arteplussept</source>
-        <news>
-v1.4.5 (2026-07-01)
-- Fix home menu display related to Arte r-proxy changes.
-- Remove auth token remote persistence.
-- Add age restriction MPAA info to videos.
-- Warn during playback if MPAA is 'PG-13', 'R' or 'NC-17'.
-v1.4.1 (2023-10-10)
-- Fix playing videos with siblings.
-v1.4.0 (2023-8-14)
-- Add support for content over multiple pages.
-- Refactor most of the code in OO style
-        </news>
+<news>1.5.1 (2026-7-15)
+
+Isolate plugin source from scripts and tests</news>
         <assets>
             <icon>resources/icon.png</icon>
             <fanart>resources/fanart.jpg</fanart>
