+++ plugin.video.arteplussept/resources/language/resource.language.pl_pl/strings.po
@@ -180,3 +180,7 @@ msgstr "Czy naprawdę chcesz wyczyścić ulubione Arte ?"
 msgctxt "#30055"
 msgid "{label} - This program contains scenes not suitable for young or sensitive audience"
 msgstr "{label} - Ten program zawiera sceny, które nie są odpowiednie dla młodych lub wrażliwych widzów"
+
+msgctxt "#30060"
+msgid "Play from the start"
+msgstr "Odtwórz od początku"
