+++ plugin.video.arteplussept/resources/language/resource.language.ro_ro/strings.po
@@ -180,3 +180,7 @@ msgstr "Doriți cu adevărat să ștergeți favoritele Arte ?"
 msgctxt "#30055"
 msgid "{label} - This program contains scenes not suitable for young or sensitive audience"
 msgstr "{label} - Programul conține scene care nu sunt potrivite pentru oamenii tineri sau sensibili"
+
+msgctxt "#30060"
+msgid "Play from the start"
+msgstr "Joacă de la început"
